requirements-dev.txt
# Development and testing dependencies
# Core testing framework
pytest>=8.4.1
pytest-django>=4.11.1

# Factory for test data generation
factory_boy>=3.3.3

# Additional testing utilities
pytest-mock>=3.12.0
pytest-cov>=5.0.0
pytest-xdist>=3.6.0

# Code quality and linting
flake8>=7.1.0
black>=24.0.0
isort>=5.13.0
autoflake>=2.3.0

# Development utilities
ipython>=8.0.0
django-debug-toolbar>=4.4.0
django-extensions>=3.2.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=2.0.0